requirements.txt
attrs==23.1.0
cattrs==23.2.3
hydra-core==1.3.2
portalocker==2.8.2
spacy==3.7.2
torch==2.1.1
torchtext==0.16.1
torchvision==0.16.1

# Non-production dependencies
pre-commit==3.6.0
pytest==7.4.3
